Moved the metric_services class to a more specific directory
There will be a lot of metrics so it helps to place all the metrics together to be more convenient.
